Error handling and securing endpoint
