+++ dev/joymeter-webapi/src/main/java/com/joymeter/exception/ErrorCode.java
@@ -2,6 +2,7 @@ package com.joymeter.exception;
 
 public enum ErrorCode {
 
+	UNAUTHORIZED(99, "Unauthorized"),
 	INVALID_FACEBOOK_TOKEN(100,"Invalid Facebook Access token"),
 	EXPIRED_FACEBOOK_TOKEN(101,"Expired Facebook Access token");
