+++ dev/joymeter-webapi/src/main/resources/joymeter.properties
@@ -1,6 +1,6 @@
 jdbc.url=jdbc:mysql://localhost:3306/joymeterdb
 jdbc.username=root
-jdbc.password=
+jdbc.password=root
 jdbc.driver=com.mysql.jdbc.Driver
 hibernate.dialect=org.hibernate.dialect.MySQLDialect
 hibernate.hbm2ddl.auto=create
