+++ dev/joymeter-webapi/src/main/webapp/WEB-INF/web.xml
@@ -14,9 +14,13 @@ xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j
         <listener-class>org.springframework.web.context.request.RequestContextListener</listener-class>
     </listener>
     <servlet>
-        <servlet-name>Jersey Spring Web Application</servlet-name>
-        <servlet-class>com.sun.jersey.spi.spring.container.servlet.SpringServlet</servlet-class>
-    </servlet>
+		<servlet-name>Jersey Spring Web Application</servlet-name>
+		<servlet-class>com.sun.jersey.spi.spring.container.servlet.SpringServlet</servlet-class>
+		<init-param>
+			<param-name>com.sun.jersey.spi.container.ContainerRequestFilters</param-name>
+			<param-value>com.joymeter.security.JoymeterContextFilter</param-value>
+		</init-param>
+	</servlet>
     <servlet-mapping>
         <servlet-name>Jersey Spring Web Application</servlet-name>
         <url-pattern>/webresources/*</url-pattern>
